refactor(cognition): fold 4 remaining perception predicates into perception.rs (pure code-motion)
Moved has_real_action_receipt (kept pub(crate)), claimed_file_without_act,
wrote_without_observation, and mutated_workspace (now pub(super)) verbatim out of mod.rs and
into the existing perception.rs, alongside the structural-fact detectors already there. mod.rs
imports them via the extended `use perception::{...}` list; no external caller changes (no
crate-external user of has_real_action_receipt exists). Also dropped a now-dead
WM_SETTLEMENT_PREFIX import in perception.rs (all uses are full-path). Behavior-identical.
mod.rs 2778 → 2650 lines; perception.rs 127 → 279 lines. act_observe tests green: 28 passed, 0 failed.
